fix: console.log double cast for %g format, STR_LIT_RUNTIME .data fix — 11 runtime failures fixed (1345 no-gcc, 1286 gcc)

+++ test/cases/phase0/console/log-f64/expected.c
@@ -2,6 +2,6 @@
 
 int main(void) {
     TSC_INIT();
-    printf("%g\n", 3.14);
+    printf("%g\n", (double)(3.14));
     return 0;
 }
