+++ test/cases/phase0/math/rounding/expected.c
@@ -3,11 +3,11 @@
 
 int main(void) {
     TSC_INIT();
-    printf("%g\n", floor(2.7));
-    printf("%g\n", ceil(2.1));
-    printf("%g\n", round(2.5));
-    printf("%g\n", round(2.4));
-    printf("%g\n", trunc(2.9));
-    printf("%g\n", trunc(-2.9));
+    printf("%g\n", (double)(floor(2.7)));
+    printf("%g\n", (double)(ceil(2.1)));
+    printf("%g\n", (double)(round(2.5)));
+    printf("%g\n", (double)(round(2.4)));
+    printf("%g\n", (double)(trunc(2.9)));
+    printf("%g\n", (double)(trunc(-2.9)));
     return 0;
 }
