+++ test/cases/phase1/arithmetic/precedence-parens/expected.c
@@ -2,6 +2,6 @@
 
 int main(void) {
     TSC_INIT();
-    printf("%g\n", (2 + 3) * 4);
+    printf("%g\n", (double)((2 + 3) * 4));
     return 0;
 }
