+++ test/cases/phase1/const/infer/expected.c
@@ -3,6 +3,6 @@
 int main(void) {
     TSC_INIT();
     const double x = 100.0;
-    printf("%g\n", x);
+    printf("%g\n", (double)(x));
     return 0;
 }
