+++ test/cases/phase1/let/f64-zero/expected.c
@@ -3,6 +3,6 @@
 int main(void) {
     TSC_INIT();
     double x = 0.0;
-    printf("%g\n", x);
+    printf("%g\n", (double)(x));
     return 0;
 }
