+++ test/cases/phase1/literals/float-point/expected.c
@@ -2,8 +2,8 @@
 
 int main(void) {
     TSC_INIT();
-    printf("%g\n", 3.14);
-    printf("%g\n", 0.5);
-    printf("%g\n", 1.0);
+    printf("%g\n", (double)(3.14));
+    printf("%g\n", (double)(0.5));
+    printf("%g\n", (double)(1.0));
     return 0;
 }
