+++ test/cases/phase1/precedence/right-assoc-pow/expected.c
@@ -3,6 +3,6 @@
 
 int main(void) {
     TSC_INIT();
-    printf("%g\n", pow(2.0, pow(3.0, 2.0)));
+    printf("%g\n", (double)(pow(2.0, pow(3.0, 2.0))));
     return 0;
 }
