+++ test/cases/phase12/math/abs/expected.c
@@ -3,6 +3,6 @@
 
 int main(void) {
     TSC_INIT();
-    printf("%g\n", fabs(-5));
+    printf("%g\n", (double)(fabs(-5)));
     return 0;
 }
