+++ test/cases/phase12/math/c-output/expected.c
@@ -4,6 +4,6 @@
 int main(void) {
     TSC_INIT();
     const double x = sqrt(16.0);
-    printf("%g\n", x);
+    printf("%g\n", (double)(x));
     return 0;
 }
