+++ test/cases/phase12/math/ceil/expected.c
@@ -3,6 +3,6 @@
 
 int main(void) {
     TSC_INIT();
-    printf("%g\n", ceil(3.2));
+    printf("%g\n", (double)(ceil(3.2)));
     return 0;
 }
