+++ test/cases/phase12/math/floor/expected.c
@@ -3,6 +3,6 @@
 
 int main(void) {
     TSC_INIT();
-    printf("%g\n", floor(3.7));
+    printf("%g\n", (double)(floor(3.7)));
     return 0;
 }
