+++ test/cases/phase12/math/min/expected.c
@@ -3,7 +3,7 @@
 
 int main(void) {
     TSC_INIT();
-    printf("%g\n", fmin(3, 5));
-    printf("%g\n", fmax(3, 5));
+    printf("%g\n", (double)(fmin(3, 5)));
+    printf("%g\n", (double)(fmax(3, 5)));
     return 0;
 }
