+++ test/cases/phase2/tuples/spread-copy/expected.c
@@ -6,7 +6,7 @@ int main(void) {
     TSC_INIT();
     const P p = {._0 = 1.0, ._1 = 2.0, ._2 = 3.0};
     const P copy = {._0 = p._0, ._1 = p._1, ._2 = p._2};
-    printf("%g\n", copy._0);
-    printf("%g\n", copy._2);
+    printf("%g\n", (double)(copy._0));
+    printf("%g\n", (double)(copy._2));
     return 0;
 }
