+++ test/cases/phase2/unknown/as-cast-class/expected.c
@@ -47,8 +47,8 @@ int main(void) {
     tsc_unknown x = tsc_unknown_from_Point(p);
     if (x.type_id == 8) {
         const Point obj = *tsc_unknown_get_Point(&x);
-        printf("%g\n", obj.x);
-        printf("%g\n", obj.y);
+        printf("%g\n", (double)(obj.x));
+        printf("%g\n", (double)(obj.y));
     }
     tsc_unknown_drop(&x);
     return 0;
