+++ test/cases/phase2/unknown/multi-type-check/expected.c
@@ -35,7 +35,7 @@ static inline void tsc_unknown_drop(tsc_unknown *self) { if (self->vtable && sel
 
 void check_unknown(tsc_unknown val) {
     if (val.type_id == 1) {
-        printf("%g\n", tsc_unknown_get_i32(&val) + 1);
+        printf("%g\n", (double)(tsc_unknown_get_i32(&val) + 1));
     }
     if (val.type_id == 6) {
         String _tmp_0 = tsc_string_to_upper(tsc_unknown_get_string(&val));
@@ -43,7 +43,7 @@ void check_unknown(tsc_unknown val) {
         tsc_string_release(_tmp_0);
     }
     if (val.type_id == 4) {
-        printf("%g\n", tsc_unknown_get_f64(&val) * 2);
+        printf("%g\n", (double)(tsc_unknown_get_f64(&val) * 2));
     }
 }
