+++ test/cases/phase2/widening/u32-to-f64/expected.c
@@ -4,6 +4,6 @@ int main(void) {
     TSC_INIT();
     const uint32_t a = 42U;
     const double b = a;
-    printf("%g\n", b);
+    printf("%g\n", (double)(b));
     return 0;
 }
