+++ test/cases/phase3/arrays/splice/expected.c
@@ -8,7 +8,7 @@ int main(void) {
     Array_f64 arr = (Array_f64){.data = _arr_data_0, .length = 5, .capacity = 5};
     const Array_f64 removed = tsc_array_splice_f64(&arr, 1, 2, 10, 20);
     printf("%zu\n", arr.length);
-    printf("%g\n", arr.data[1]);
+    printf("%g\n", (double)(arr.data[1]));
     printf("%zu\n", removed.length);
     return 0;
 }
