+++ test/cases/phase3/arrays/to-reversed/expected.c
@@ -7,7 +7,7 @@ int main(void) {
     double _arr_data_0[] = {1, 2, 3};
     const Array_f64 arr = (Array_f64){.data = _arr_data_0, .length = 3, .capacity = 3};
     const Array_f64 rev = tsc_array_to_reversed_f64(arr);
-    printf("%g\n", rev.data[0]);
-    printf("%g\n", arr.data[0]);
+    printf("%g\n", (double)(rev.data[0]));
+    printf("%g\n", (double)(arr.data[0]));
     return 0;
 }
