+++ test/cases/phase3/ownership/tuple-primitive-copy/expected.c
@@ -7,7 +7,7 @@ int main(void) {
     tuple_i32_f64 pair = {._0 = 1, ._1 = 2.0};
     tuple_i32_f64 b = pair;
     printf("%d\n", b._0);
-    printf("%g\n", b._1);
+    printf("%g\n", (double)(b._1));
     printf("%d\n", pair._0);
     return 0;
 }
