+++ test/cases/phase6/interop/declare-function/expected.c
@@ -5,6 +5,6 @@ extern double sin(double x);
 
 int main(void) {
     TSC_INIT();
-    printf("%g\n", sin(0.0));
+    printf("%g\n", (double)(sin(0.0)));
     return 0;
 }
